\name{venn.plot}
\alias{venn.plot}
\title{A collection of venn diagram grob objects}
\description{Examples of venn.diagram. For testing that the venn.diagrams are generated properly}
\value{Depends on which file is loaded. Can be a list of one to five category Venn Diagrams}
\author{Zhiyuan Wang}
\keyword{ internal }
