\name{VennDiagram-internal}

%% Internal functions not intended to be exposed.  Some of these are
%% potentially useful/generalizable grid utility functions that might
%% be exported in the future.
%% <LIST>
\alias{adjust.venn}
\alias{circle}
\alias{add.title}
\alias{decide.special.case}
\alias{draw.001}
\alias{draw.010}
\alias{draw.011A}
\alias{draw.011O}
\alias{draw.012AA}
\alias{draw.021AA}
\alias{draw.022AAAO}
\alias{draw.022AAOO}
\alias{draw.023}
\alias{draw.032}
\alias{draw.033}
\alias{draw.100}
\alias{draw.110}
\alias{draw.111A}
\alias{draw.112AA}
\alias{draw.120}
\alias{draw.121AO}
\alias{draw.122AAOO}
\alias{draw.130}
\alias{draw.pairwise.venn}
\alias{draw.quad.venn}
\alias{draw.single.venn}
\alias{draw.sp.case}
\alias{draw.triple.venn}
\alias{ell2poly}
\alias{ellipse}
\alias{find.cat.pos}
\alias{find.dist}
\alias{find.intersect}
\alias{flip.venn}
\alias{rotate}
\alias{rotate.sp}
\alias{rotate.venn.degrees}
%% </LIST>


\title{Internal VennDiagram Functions}
\description{
  Internal VennDiagram functions
}
\details{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written :).
}
\keyword{ internal }
