\name{VennDiagram-package}
\alias{VennDiagram-package}
\alias{VennDiagram}
\docType{package}
\title{Venn diagram plotting}
\description{Functions to plot high-resolution and highly-customizable Venn and Euler plots.}
\details{
\tabular{ll}{
Package: \tab VennDiagram\cr
Type: \tab Package\cr
Version: \tab 1.6.0\cr
Date: \tab 2013-04-10\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}
\author{
Author: Hanbo Chen <Hanbert.Chen@mail.utoronto.ca>\cr
Maintainer: Dr. Paul C. Boutros <Paul.Boutros@utoronto.ca>
}
\keyword{ package }
