\name{VennDiagram-internal}

%% Internal functions not intended to be exposed.  Some of these are
%% potentially useful/generalizable grid utility functions that might
%% be exported in the future.
%% <LIST>
\alias{adjust.venn}
\alias{circle}
\alias{add.title}
\alias{decide.special.case}
\alias{draw.sp.case}
\alias{ell2poly}
\alias{ellipse}
\alias{find.cat.pos}
\alias{find.dist}
\alias{find.intersect}
\alias{flip.venn}
\alias{rotate}
\alias{rotate.sp}
\alias{rotate.venn.degrees}
%% </LIST>


\title{Internal VennDiagram Functions}
\description{
  Internal VennDiagram functions
}
\details{
  These are not to be called by the user (or in some cases are just
  waiting for proper documentation to be written :)
}
\keyword{ internal }
